third_party/blink/web_tests/platform/linux-chrome/external/wpt/web-share/disabled-by-permissions-policy-cross-origin.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] share() is disabled by default 'self' by permissions policy for cross-origin iframes
  assert_equals: expected "NotAllowedError" but got "Error"
[FAIL] share() is disabled explicitly by permissions policy for cross-origin iframe
  assert_equals: expected "NotAllowedError" but got "Error"
[FAIL] share() not allowed, as only allowed to share with self
  assert_equals: expected "NotAllowedError" but got "Error"
[PASS] canShare() not allowed to share by default permissions policy cross-origin
[PASS] canShare() is allowed by permissions policy to share cross-origin on a particular origin
[PASS] canShare() with self
Harness: the test ran to completion.
